src/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
certifi==2021.5.30
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-location-field==2.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dropbox==11.30.0
idna==3.3
Markdown==3.3.7
MarkupPy==1.14
Pillow==9.1.1
ply==3.11
promise==2.3
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
python-memcached==1.59
pytz==2021.1
PyYAML==5.4.1
requests==2.27.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
stone==3.2.1
tablib==3.0.0
text-unidecode==1.3
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
